+++ SniffCore.Layouting/CloseButton.cs
@@ -1,0 +1,33 @@
+﻿using System.ComponentModel;
+using System.Windows;
+using System.Windows.Controls;
+
+namespace SniffCore.Layouting
+{
+    /// <summary>
+    ///     Represents the close Button shown in the <see cref="DynamicTabControl" />.
+    /// </summary>
+    public class CloseButton : Button
+    {
+        /// <summary>
+        ///     Identifies the <see cref="StrokeThickness" /> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty StrokeThicknessProperty =
+            DependencyProperty.Register("StrokeThickness", typeof(double), typeof(CloseButton), new UIPropertyMetadata(1.5));
+
+        static CloseButton()
+        {
+            DefaultStyleKeyProperty.OverrideMetadata(typeof(CloseButton), new FrameworkPropertyMetadata(typeof(CloseButton)));
+        }
+
+        /// <summary>
+        ///     Gets or sets the stroke thickness of the X in the template.
+        /// </summary>
+        [DefaultValue(1.5)]
+        public double StrokeThickness
+        {
+            get => (double) GetValue(StrokeThicknessProperty);
+            set => SetValue(StrokeThicknessProperty, value);
+        }
+    }
+}