+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Ideas/Upcoming]
 * Outlookbar (NavigationBar, Accordion)
 * Pageinator (PagingControl)
-* DynamicTabControl
 
 ## [Unreleased]
 * Added ArcPanel which arranges child elements in a configurable arc form.
@@ -14,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Added UniformWrapPanel which enhances the WrapPanel by the feature that all items will have the same size.
 * Added Resizer which brings the possibility to resize every UI control manually by hold and drag the corners or sides.
 * Added HeaderItemsControl to allign items with a header and content vertically or horizontally.
+* Added a DynamicTabControl which is a tab control with add and remove buttons.
 
 ## [1.0.0] - 2021-06-13
 ### Added